Compute atmospheric pressure from location altitude if coded with special value in landscape.ini.

+++ src/core/modules/Landscape.hpp
@@ -77,9 +77,10 @@ public:
     	const int getDefaultFogSetting() const {return defaultFogSetting;}
   	//! Return default atmosperic extinction, mag/airmass, or -1 (no change)
     	const float getDefaultAtmosphericExtinction() const {return defaultExtinctionCoefficient;}
-        //! Return default atmospheric temperature, for refraction computation.
+        //! Return default atmospheric temperature, for refraction computation, or -1000 for "unknown/no change".
         const float getDefaultAtmosphericTemperature() const {return defaultTemperature;}
         //! Return default atmospheric temperature, for refraction computation.
+        //! returns -1 to signal "standard conditions", or -2 for "unknown/invalid/no change"
         const float getDefaultAtmosphericPressure() const {return defaultPressure;}
 
 	//! Set the z-axis rotation (offset from original value when rotated
@@ -109,9 +110,9 @@ protected:
 	int cols; // vertical columns
         int defaultBortleIndex; // light pollution from landscape.ini, or -1(no change)
         int defaultFogSetting; // fog flag setting from landscape.ini: -1(no change), 0(off), 1(on)
-        float defaultExtinctionCoefficient; // atmospheric_extinction_coefficient from landscape.ini or -1
-        float defaultTemperature; // atmospheric_temperature from landscape.ini or -1000.0
-        float defaultPressure; // atmospheric_pressure from landscape.ini or -1.0
+        double defaultExtinctionCoefficient; // atmospheric_extinction_coefficient from landscape.ini or -1
+        double defaultTemperature; // atmospheric_temperature from landscape.ini or -1000.0
+        double defaultPressure; // atmospheric_pressure from landscape.ini or -1.0
 
 	typedef struct
 	{
